refactor: Button, CircularProgress 컴포넌트 폴더 이동
components common으로 이동
