+++ frontend/src/components/common/Button/Button.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import Button from './Button';
 import { css } from 'styled-components';
-import color from '../../styles/color';
+import color from '../../../styles/color';
 
 type Story = StoryObj<typeof Button>;
