+++ frontend/src/components/common/CircularProgress/CircularProgress.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import CircularProgress from './CircularProgress';
 import { css } from 'styled-components';
-import color from '../../styles/color';
+import color from '../../../styles/color';
 
 type Story = StoryObj<typeof CircularProgress>;
